+++ src/cloudant/_common_util.py
@@ -283,6 +283,8 @@ class _Code(str):
     codifying map and reduce Javascript content.
     """
     def __new__(cls, code):
+        if isinstance(code, unicode):
+            return str.__new__(cls, code.encode('utf8'))
         return str.__new__(cls, code)
 
 class InfiniteSession(Session):