+++ CHANGES.rst
@@ -3,6 +3,8 @@
 - [FIXED] Fixed ``TypeError`` when setting revision limits on Python>=3.6.
 - [FIXED] Fixed the ``exists()`` double check on ``client.py`` and ``database.py``.
 - [FIXED] Fixed Cloudant exception code 409 with 412 when creating a database that already exists.
+- [FIXED] Catch error if ``throw_on_exists`` flag is ``False`` for document create.
+- [FIXED] Fixed /_all_docs call where ``keys`` is an empty list.
 
 2.4.0 (2017-02-14)
 ==================