+++ CHANGES.rst
@@ -1,5 +1,6 @@
 2.5.0 (Unreleased)
 ==================
+- [FIXED] Fixed crash caused by non-UTF8 chars in ddocs
 - [FIXED] Fixed ``TypeError`` when setting revision limits on Python>=3.6.
 - [FIXED] Fixed the ``exists()`` double check on ``client.py`` and ``database.py``.
 - [FIXED] Fixed Cloudant exception code 409 with 412 when creating a database that already exists.